config.json: A new field "default_command_icons" is added, which is another json object of keys -> url mappings, each representing a command name and the command's respective icon. These are always default to be used on a command thumbnail unless a server specifies a different icon for a command. On each server, a new field "command_icons" is added, which is another json object of keys -> url mappings, each representing a command name and the command's respective icon. On each server, a new field "static_status" *may* be added, which denotes that on the `status` command, this will be displayed in place of a ping to the JQM server.
MediusJQMServer.java:
Added a new field `private String staticStatus;`, which denotes that on the `status` command, this will be displayed in place of a ping to the JQM server.
Added a new field `private final HashMap<String,String> commandIcons;`, which is a key, value pair where the key is the name of the command(String), and the value is a url to the respective command icon (String). The primary url source for a particular server's icon, if it exists.

MediusBotConfig.java:
Added a new field `private HashMap<String, String> defaultCommandIcons;`, which is a key, value pair where the key is the name of the command(String), and the value is a url to the respective command icon (String). Used by default if command is not server spcified, and as a fallback if a server does not have a specified icon for a command.

When parsing servers, we handle the parsing of "default_command_icons", creating a new hash map to populate the mappings of default command name to command icon url.
When parsing servers, we handle the parsing of "command_icons", creating a new hash map to populate the mappings of server specific command name to command icon url.

HelpCommand.java:
Revert the output of commands to what it originally was; printing first non operator commands, and then operator commands.
Remove extra lines, and remove footer.

StatusCommand.java:
Remove new declarations of hashmaps for status/ thumbnails.
Directly load icons from the config default/ server specified command icons.
Check if we have a specified static message already pre-loaded in config.

Add private method `getStatus(MediusJQMServer server)` to retrieve the status of the specified medius server.